feat: add core config module with tier mapping
